Revert "x86_64: fix LP64 format/cast errors in drivers and servers, add amd64 distribution sets"
This reverts commit cdc3109e2497f0c389e60c994bcc1b12e1bfbad0.

+++ minix/drivers/storage/fbd/fbd.c
@@ -17,7 +17,7 @@
 /* Function declarations. */
 static int fbd_open(devminor_t minor, int access);
 static int fbd_close(devminor_t minor);
-static ssize_t fbd_transfer(devminor_t minor, int do_write, u64_t position,
+static int fbd_transfer(devminor_t minor, int do_write, u64_t position,
 	endpoint_t endpt, iovec_t *iov, unsigned int nr_req, int flags);
 static int fbd_ioctl(devminor_t minor, unsigned long request, endpoint_t endpt,
 	cp_grant_id_t grant, endpoint_t user_endpt);
@@ -289,7 +289,7 @@ static ssize_t fbd_transfer_copy(int do_write, u64_t position,
 	assert(count > 0 && count <= SCPVEC_NR);
 
 	if (size > BUF_SIZE) {
-		printf("FBD: allocating memory for %zu bytes\n", size);
+		printf("FBD: allocating memory for %d bytes\n", size);
 
 		ptr = alloc_contig(size, 0, NULL);
 
@@ -393,7 +393,7 @@ static ssize_t fbd_transfer_copy(int do_write, u64_t position,
 /*===========================================================================*
  *				fbd_transfer				     *
  *===========================================================================*/
-static ssize_t fbd_transfer(devminor_t UNUSED(minor), int do_write, u64_t position,
+static int fbd_transfer(devminor_t UNUSED(minor), int do_write, u64_t position,
 	endpoint_t endpt, iovec_t *iov, unsigned int nr_req, int flags)
 {
 	/* Transfer data from or to the device. */
